+++ src/query/test/compatibility/testdata/aggregators.test
