+++ src/query/test/compatibility/testdata/functions.test
