+++ src/query/test/compatibility/testdata/histograms.test
