+++ src/query/test/compatibility/testdata/legacy.test
