+++ src/query/test/compatibility/testdata/literals.test
