+++ src/query/test/compatibility/testdata/operators.test
