+++ src/query/test/compatibility/testdata/selectors.test
