+++ src/query/test/compatibility/testdata/subquery.test
